requirements.txt
# Add your project dependencies here
telethon
python-dotenv